feat: Move private components to root-level dir
