+++ ember-toucan-core/src/-private/components/hint.hbs
