+++ ember-toucan-core/src/-private/components/label.hbs
